backend/requirements.txt
# Core dependencies for the backend API
fastapi==0.120.0
uvicorn==0.38.0
python-dotenv==1.1.1
python-multipart==0.0.20
anthropic==0.71.0

# Supabase client
supabase==2.22.2

# Data processing and analysis
pandas==2.3.3
numpy>=1.26.0
yfinance==0.2.66
matplotlib>=3.0.0

# HTTP client
requests>=2.31.0
